Dynamic background star speed
Background stars now move at increasing pace as player progresses in
game. See Constants.as for the values affecting this.
(Note that the current random expression we use for spawning stars is a
bit strange in that it depends not only on game time, but also on the
rate at which we call updateStarSpawning()... we might want to change
this if we want more uniform star density as the speed of stars
increases)